requirements.txt
django==5.1
djangorestframework==3.15.2
pytest-django==4.9.0
Faker==33.0.0